feat: finish authenticated routes in server routes

+++ server/src/routes/memories.ts
@@ -3,6 +3,10 @@ import { z } from 'zod'
 import { prisma } from '../lib/prisma'
 
 export async function memoriesRoute(app: FastifyInstance) {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
+
   app.get('/memories', async () => {
     const memories = await prisma.memory.findMany({
       orderBy: {
@@ -17,7 +21,7 @@ export async function memoriesRoute(app: FastifyInstance) {
     }))
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -29,6 +33,10 @@ export async function memoriesRoute(app: FastifyInstance) {
         id,
       },
     })
+
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     return memory
   })
@@ -47,14 +55,14 @@ export async function memoriesRoute(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '1',
+        userId: request.user.sub,
       },
     })
 
     return memory
   })
 
-  app.put('/memories', async (request) => {
+  app.put('/memories', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -69,7 +77,17 @@ export async function memoriesRoute(app: FastifyInstance) {
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -83,12 +101,22 @@ export async function memoriesRoute(app: FastifyInstance) {
     return memory
   })
 
-  app.delete('/memories', async (request) => {
+  app.delete('/memories', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
